event trace plumbing for multi-process support
related review:
http://codereview.chromium.org/6691013/

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6694004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@78768 0039d316-1c4b-4281-b951-d872f2087c98

+++ gpu/common/gpu_trace_event.cc
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2010 The Chromium Authors. All rights reserved.
+// Copyright (c) 2011 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -130,7 +130,7 @@ TraceLog::~TraceLog() {
 TraceCategory* TraceLog::GetCategory(const char* name) {
   AutoLock lock(lock_);
   // TODO(nduca): replace with a hash_map.
-  for (int i = categories_.size() - 1; i >= 0; i-- ) {
+  for (int i = static_cast<int>(categories_.size()) - 1; i >= 0; i-- ) {
     if (strcmp(categories_[i]->name(), name) == 0)
       return categories_[i];
   }
